- Refactored web install code in Cmake - Install now support BUILD_NO_OAUTH options for isolated testing - Basic auth works, globsu auth not tested yet
