+++ scripts/install_ws_service.sh
@@ -7,22 +7,23 @@ SOURCE=$(dirname "$SCRIPT")
 PROJECT_ROOT=$(realpath ${SOURCE}/..)
 
 # Make sure paths exist
-mkdir -p /opt/datafed/web
-mkdir -p /opt/datafed/keys
-mkdir -p /var/log/datafed
+#mkdir -p /opt/datafed/web
+#mkdir -p /opt/datafed/keys
+#mkdir -p /var/log/datafed
 
 # Install web node modules
-cp "$PROJECT_ROOT/web/package.json" /opt/datafed/web/
+#cp "$PROJECT_ROOT/web/package.json" /opt/datafed/web/
 #cp "$PROJECT_ROOT/web/package-lock.json" /opt/datafed/web/
 
-export npm_config_cache=/opt/datafed/web
-npm --allow-root --unsafe-perm --prefix /opt/datafed/web install 
+#export npm_config_cache=/opt/datafed/web
+#npm --allow-root --unsafe-perm --prefix /opt/datafed/web install 
+
 # Install javascript web server repo and core server were 
 # already installed by CMake
-cp "$PROJECT_ROOT/web/datafed-ws.js" /opt/datafed/web
+#cp "$PROJECT_ROOT/web/datafed-ws.js" /opt/datafed/web
 
 # Copy configuration files
-cp "$PROJECT_ROOT/config/datafed-ws.cfg" /opt/datafed/web
+#cp "$PROJECT_ROOT/config/datafed-ws.cfg" /opt/datafed/web
 
 # Copy services
 cp "$PROJECT_ROOT/services/datafed-ws.service" /etc/systemd/system
