+++ web/auth/static/auth_globus_register.js
